requirements.txt
python-dotenv
groq
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
python-docx
openpyxl